requirements.txt
fastapi
uvicorn 
numpy
matplotlib
opencv-python-headless
scikit-learn
Pillow
pillow-heif
python-multipart 